leaflet layer fixing
